Prevent animal pic deletion on edit
Add hidden input for animal pic in order to always get an animal pic on
editing, this is to prevent a the case a user edits an animal and do not changes the animal
photo input. The input will come as empty and hence updates the animal with no
picture at all.